+++ examples/webpack.common.js
@@ -25,6 +25,6 @@ module.exports = {
     },
     performance: {
         hints: 'error',
-        maxAssetSize: 4.6 * mb,
+        maxAssetSize: 4.7 * mb,
     },
 };
