+++ lib/core/src/control.rs
@@ -2,8 +2,6 @@
 
 pub mod callback;
 pub mod event_loop;
-pub mod frp;
-pub mod frp2;
 pub mod io;
 
 pub use event_loop::*;
